Fixed an issue introduced in r2626 (I guess) where Lua strings would magically become true (for instance, in: get, triggerServerEvent and triggerClientEvent)
Apparently C++ became confused because there wasn't a char* overloaded member anymore, therefore it decided to go with the bool overloaded member. This commits hints to converting it to a std::string and use that overloaded member instead.

Nice of C++ to just drop a single warning message only once and not to mention it ever again after that.

(The fix in CGame.cpp is nothing interesting, just a warning fix for Linux, whining about && and || (and it just looks nicer))