[CoreVoice] Added microphone menu. - Current options are enable local playback (hear yourself), a VU meter and device selection : http://prntscr.com/537mzm - TODO:   - Take into account unicode spacing issues with Stop Testing/Test buttons and non English languages   - Add new images to the data files pack - need to find out where to do this from ccw as they aren't in the main source tree   - Add quality labels once the volume levels for each bar are finalised   - Look into some crackling on local playback, this could be a hardware issue   - Update BASS libraries to fix some unrelated multi threading issues we might be having
Lastly I just want to point out test mode is local testing only, it's not sent to the server you are connected to (if you are)

p.s.
for any developers reading I also added a current tab field to CSettings so that I didn't have to update anything not active. the joypad tab used to update every 30 frames even if you don't have the settings menu open.